nix-shell: Add keymap-drawer and flash packages
## Changes Description

This commit adds the `keymap-drawer` utility to the dev shell so we can
generate layout map images of our configuration. It also adds the
`flash` package from the flake as a binary, so one can simply call
`zmk-uf2-flash` inside a dev shell to flash the firmware built by the
flake.